\name{PhytosanitaryCalculator-package}
\alias{PhytosanitaryCalculator-package}
\alias{PhytosanitaryCalculator}
\docType{package}
\title{Phytosanitary Calculator}
\description{
A Shiny application for calculating phytosanitary inspection plans based on risks.
}
\details{
The application provides sampling plans with minimal sample sizes that keep both producer and importer risks below user-defined levels.
}
\author{
Your Name
Maintainer: Your Name <your.email@example.com>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
