% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phitosanitarycalculator.R
\name{run_app}
\alias{run_app}
\title{Phytosanitary Calculator}
\usage{
run_app()
}
\value{
A Shiny app object.
}
\description{
A Shiny application for calculating phytosanitary inspection plans based on risks.
}
\details{
The application provides sampling plans with minimal sample sizes that keep both producer and importer risks below user-defined levels.
}
