\name{PhytosanitaryCalculator-package}
\alias{PhytosanitaryCalculator-package}
\alias{PhytosanitaryCalculator}
\docType{package}
\title{
Phytosanitary Calculator for Inspection Plans Based on Risks
}
\description{
A Shiny application for calculating phytosanitary inspection plans based on risks.
}
\details{
The application provides sampling plans with minimal sample sizes that keep both producer and importer risks below user-defined levels.
}
\author{
Gustavo Ramirez-Valverde, Luis Gabriel Otero-Prevost, Pedro Macias-Canales, Juan A. Villanueva-Jimenez, Jorge Luis Leyva-Vazquez
Maintainer: Gustavo Ramirez-Valverde <gustavoramirezvalverde@gmail.com>
}
\references{
# Incluye aquí tus referencias relevantes
Guenther, W. C. (1977) Sampling Inspection in statistical quality control, Charles Griffin and Company Limited
Hald, A. (1981), Statistical theory of sampling inspection by attributes, Academic Press.
Schilling, E. G. (1982), Acceptance Sampling in Quality Control, Dekker

}
\examples{
if(interactive()) {
  run_app()
}
}
\keyword{package}
